paper4/requirements.txt
# Paper 4: Alignment Robustness - Pinned Requirements
# Tested configuration for reproducibility
# Created: 2026-01-20

# Core ML
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0
bitsandbytes==0.41.0

# Scientific computing
scipy==1.11.0
numpy==1.24.0
pandas==2.0.0

# Configuration and validation
pyyaml==6.0
jsonschema==4.20.0

# Jupyter notebooks
jupyter==1.0.0
ipywidgets==8.1.0

# Optional: for plots
matplotlib==3.8.0
seaborn==0.13.0
